requirements.txt
alabaster==0.7.12
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
docutils==0.16
dropbox==9.4.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
gspread==3.0.1
gspread-dataframe==3.0.2
httplib2==0.17.3
idna==2.7
imagesize==1.2.0
importlib-metadata==1.6.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
more-itertools==8.3.0
numpy==1.14.3
oauth2client==4.1.2
oauthlib==3.1.0
packaging==20.4
pandas==0.23.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.7.3
pytz==2020.1
PyYAML==5.3.1
requests==2.19.1
requests-oauthlib==1.2.0
rsa==4.0
six==1.14.0
slackclient==1.3.0
Sphinx==3.0.4
SQLAlchemy==1.3.17
traitlets==4.3.3
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
validators==0.13.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
zipp==3.1.0
